requirements.txt
Pillow
Flask 